added fournisseurs / achats /caisse / fixed vente /historique  bugs added login forms i had a problem in Pushing previous commits so i had to delete all and add new commit

+++ DataBase/Models/Achat.cs
@@ -9,7 +9,7 @@ namespace Gestionnaire_Pro.DataBase.Models
     {
         public int Id { get; set; }
         public DateTime dateAchat { get; set; }
-        public decimal montantTotal { get; set; }
+        public float montantTotal { get; set; }
 
         public ICollection<DetailAchat> DetailAchats { get; set; } = new List<DetailAchat>();
     }
